+++ packages/app/src/app/pages/Explore/elements.js
@@ -1,7 +1,5 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import Title from 'app/components/Title';
-import NextIcon from 'react-icons/lib/md/navigate-next';
-import PrevIcon from 'react-icons/lib/md/navigate-before';
 
 export const Heading = styled(Title)`
   margin-top: 7rem;
@@ -29,47 +27,4 @@ export const FancyHeader = styled.section`
   border-radius: 10% 10% 50% 50% / 0% 0% 31% 34%;
   box-shadow: 0 0 14px 15px rgba(0, 0, 0, 0.08);
   background: #48b2f5;
-`;
-
-export const ModalContainer = styled.div`
-  background-color: ${props => props.theme.background};
-  padding: 1rem;
-  padding-top: 0.75rem;
-
-  margin: 0;
-  color: rgba(255, 255, 255, 0.8);
-`;
-
-const Icon = css`
-  width: 50px;
-  height: 50px;
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  cursor: pointer;
-`;
-
-export const NextIconStyled = styled(NextIcon)`
-  ${Icon};
-  right: -50px;
-`;
-export const PrevIconStyled = styled(PrevIcon)`
-  ${Icon};
-  left: -50px;
-`;
-
-export const Content = styled.section`
-  margin-top: 3rem;
-  display: grid;
-  grid-template-columns: 2fr 1fr;
-  grid-column-gap: 60px;
-`;
-
-export const Tag = styled.span`
-  position: relative;
-  color: white;
-  background-color: ${props => props.theme.secondary};
-  padding: 0.3em 0.5em;
-  border-radius: 4px;
-  font-weight: 500;
 `;
